refactor: move Home container to directly under page folder
