+++ src/containers/index.tsx
@@ -3,7 +3,7 @@ import {
   Switch,
   Route,
 } from 'react-router-dom';
-import Home from 'containers/pages/home/index';
+import Home from 'containers/pages/index';
 import { VFC } from 'react';
 
 const Containers: VFC = () => (
